requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
streamlit>=1.29.0
chromadb>=0.4.18
cryptography>=41.0.0
python-dotenv>=1.0.0
ollama>=0.1.6
numpy>=1.24.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0
langgraph>=0.0.15
sentence-transformers>=2.2.2
# docling
# dotenv
# setuptools
# wheel